move test webpack config out
