+++ generators/client/files-angular.js
@@ -43,6 +43,7 @@ const files = {
                 'webpack/_webpack.dev.js',
                 'webpack/_webpack.prod.js',
                 'webpack/_webpack.vendor.js',
+                'webpack/_webpack.test.js',
                 { file: 'webpack/logo-jhipster.png', method: 'copy' }
             ]
         }
