+++ generators/client/templates/angular/_package.json
@@ -132,8 +132,8 @@
     "node": ">=6.9.0"
   },
   "scripts": {
-    "lint": "tslint '<%= MAIN_SRC_DIR%>app/**/*.ts'",
-    "lint:fix": "tslint '<%= MAIN_SRC_DIR%>app/**/*.ts' --fix",
+    "lint": "tslint '<%= MAIN_SRC_DIR%>app/**/*.ts' '<%= MAIN_TEST_DIR%>**/*.ts'",
+    "lint:fix": "tslint '<%= MAIN_SRC_DIR%>app/**/*.ts' '<%= MAIN_TEST_DIR%>**/*.ts' --fix",
     "ngc": "ngc -p tsconfig-aot.json",
     "cleanup": "rimraf <%= BUILD_DIR %>{aot,www}",
     "start": "<%= clientPackageManager %> run webpack:dev",
@@ -143,7 +143,7 @@
     "webpack:build": "<%= clientPackageManager %> run cleanup && <%= clientPackageManager %> run webpack:build:vendor && <%= clientPackageManager %> run webpack:build:dev",
     "webpack:prod": "<%= clientPackageManager %> run cleanup && webpack -p --config webpack/webpack.vendor.js --progress --profile && webpack -p --config webpack/webpack.prod.js --progress --profile",
     "webpack:test": "<%= clientPackageManager %> run test",
-    "test": "<%= clientPackageManager %> run lint && karma start src/test/javascript/karma.conf.js",
+    "test": "karma start src/test/javascript/karma.conf.js",
     "test:watch": "karma start src/test/javascript/karma.conf.js --watch",
     <%_ if (protractorTests) { _%>
     "e2e": "protractor <%= TEST_SRC_DIR %>protractor.conf.js",
