+++ generators/client/templates/angular/src/test/javascript/_karma.conf.js
@@ -16,13 +16,9 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 -%>
-const path = require('path');
-const webpack = require('webpack');
-const LoaderOptionsPlugin = require('webpack/lib/LoaderOptionsPlugin');
+const webpackConfig = require('../../../webpack/webpack.test.js');
 
 const WATCH = process.argv.indexOf('--watch') > -1;
-
-const root = __path => path.join(__dirname, __path);
 
 module.exports = (config) => {
     config.set({
@@ -49,63 +45,7 @@ module.exports = (config) => {
             'spec/entry.ts': ['webpack', 'sourcemap']
         },
 
-        webpack: {
-            resolve: {
-                extensions: ['.ts', '.js']
-            },
-            module: {
-                rules: [
-                    {
-                        test: /\.ts$/, enforce: 'pre', loader: 'tslint-loader', exclude: /(test|node_modules)/
-                    },
-                    {
-                        test: /\.ts$/,
-                        loaders: ['awesome-typescript-loader', 'angular2-template-loader?keepUrl=true'],
-                        exclude: /node_modules/
-                    },
-                    {
-                        test: /\.(html|css)$/,
-                        loader: 'raw-loader',
-                        exclude: /\.async\.(html|css)$/
-                    },
-                    {
-                        test: /\.async\.(html|css)$/,
-                        loaders: ['file?name=[name].[hash].[ext]', 'extract']
-                    },
-                    <%_ if (useSass) { _%>
-                    {
-                        test: /\.scss$/,
-                        loaders: ['to-string-loader', 'css-loader', 'sass-loader']
-                    },
-                    <%_ } _%>
-                    {
-                        test: /\.(jpe?g|png|gif|svg|woff2?|ttf|eot)$/i,
-                        loaders: ['file-loader?hash=sha512&digest=hex&name=[hash].[ext]']
-                    },
-                    {
-                        test: /src[/|\\]main[/|\\]webapp[/|\\].+\.ts$/,
-                        enforce: 'post',
-                        exclude: /(test|node_modules)/,
-                        loader: 'sourcemap-istanbul-instrumenter-loader?force-sourcemap=true'
-                    }]
-            },
-            devtool: 'inline-source-map',
-            plugins: [
-                new webpack.ContextReplacementPlugin(
-                    // The (\\|\/) piece accounts for path separators in *nix and Windows
-                    /angular(\\|\/)core(\\|\/)(esm(\\|\/)src|src)(\\|\/)linker/,
-                    root('./src') // location of your src
-                ),
-                new LoaderOptionsPlugin({
-                    options: {
-                        tslint: {
-                            emitErrors: !WATCH,
-                            failOnHint: false
-                        }
-                    }
-                })
-            ]
-        },
+        webpack: webpackConfig(WATCH),
 
         // test results reporter to use
         // possible values: 'dots', 'progress'
